# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/partner/clients": {
      "get": {
        "operationId": "PartnerController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "grantId": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "workspaceId": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "name": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "trial",
                              "active",
                              "suspended",
                              "closed"
                            ]
                          },
                          "companies": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "lastActivityAt": {
                            "nullable": true,
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "grantId",
                          "workspaceId",
                          "name",
                          "status",
                          "companies",
                          "lastActivityAt",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": {
                      "nullable": true,
                      "type": "string"
                    },
                    "usage": {
                      "type": "object",
                      "properties": {
                        "clientWorkspaces": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "limit": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "clientWorkspaces",
                        "limit"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor",
                    "usage"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "partner"
        ]
      },
      "post": {
        "operationId": "PartnerController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Replays the first response for 24 h",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 120
                      },
                      "country": {
                        "type": "string",
                        "pattern": "^[A-Z]{2}$"
                      },
                      "defaultLocale": {
                        "default": "en",
                        "type": "string",
                        "pattern": "^[a-z]{2,3}(-[A-Z]{2})?$"
                      },
                      "timezone": {
                        "default": "UTC",
                        "type": "string",
                        "maxLength": 64
                      },
                      "dataRegion": {
                        "default": "eu",
                        "type": "string",
                        "enum": [
                          "eu",
                          "in",
                          "us"
                        ]
                      }
                    },
                    "required": [
                      "name",
                      "country",
                      "defaultLocale",
                      "timezone",
                      "dataRegion"
                    ],
                    "additionalProperties": false
                  },
                  "company": {
                    "type": "object",
                    "properties": {
                      "legalName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                      },
                      "displayName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 120
                      },
                      "registrationNo": {
                        "nullable": true,
                        "type": "string",
                        "maxLength": 64
                      },
                      "sectorCode": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 64,
                        "description": "Sector code from GET /reference/sectors"
                      },
                      "sizeBand": {
                        "type": "string",
                        "enum": [
                          "micro",
                          "small",
                          "medium",
                          "large"
                        ]
                      },
                      "employeeCount": {
                        "nullable": true,
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 10000000
                      },
                      "country": {
                        "type": "string",
                        "pattern": "^[A-Z]{2}$",
                        "description": "ISO 3166-1 alpha-2 country code from GET /reference/countries"
                      },
                      "region": {
                        "nullable": true,
                        "type": "string",
                        "maxLength": 120
                      },
                      "website": {
                        "nullable": true,
                        "type": "string",
                        "maxLength": 300,
                        "format": "uri"
                      },
                      "description": {
                        "nullable": true,
                        "type": "string",
                        "maxLength": 2000
                      },
                      "address": {
                        "nullable": true,
                        "type": "object",
                        "properties": {
                          "line1": {
                            "nullable": true,
                            "type": "string",
                            "maxLength": 200
                          },
                          "line2": {
                            "nullable": true,
                            "type": "string",
                            "maxLength": 200
                          },
                          "city": {
                            "nullable": true,
                            "type": "string",
                            "maxLength": 120
                          },
                          "postalCode": {
                            "nullable": true,
                            "type": "string",
                            "maxLength": 32
                          },
                          "state": {
                            "nullable": true,
                            "type": "string",
                            "maxLength": 120
                          }
                        },
                        "additionalProperties": false
                      },
                      "fiscalYearStartMonth": {
                        "default": 1,
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 12
                      },
                      "currency": {
                        "type": "string",
                        "pattern": "^[A-Z]{3}$",
                        "description": "ISO 4217 currency code"
                      },
                      "parentCompanyId": {
                        "nullable": true,
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      }
                    },
                    "required": [
                      "legalName",
                      "sectorCode",
                      "sizeBand",
                      "country",
                      "fiscalYearStartMonth",
                      "currency"
                    ],
                    "additionalProperties": false
                  },
                  "owner": {
                    "type": "object",
                    "properties": {
                      "email": {
                        "type": "string",
                        "maxLength": 254,
                        "format": "email",
                        "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      }
                    },
                    "required": [
                      "email"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "workspace",
                  "company",
                  "owner"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client workspace created; the owner was invited by email",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "workspaceId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "companyId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "grantId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "invitationId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "required": [
                    "workspaceId",
                    "companyId",
                    "grantId",
                    "invitationId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "partner"
        ]
      }
    },
...
  },
  "info": {
    "title": "ResiliSense API",
    "description": "ResiliSense 2.0 REST API. Specs: docs/revamp/modules. Errors are RFC 9457 problem+json.",
    "version": "2.0.0-alpha",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}